requirements.txt
aldryn-apphooks-config==0.6.0
aldryn-boilerplates==0.8.0
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
cachetools==4.2.2
celery==5.1.2
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
cssutils==2.3.0
dj-database-url==0.5.0
Django==3.1.11
django-absolute==0.3
django-appconf==1.0.4
django-appdata==0.3.2
django-bootstrap-modal-forms==2.2.0
django-classy-tags==2.0.0
django-cms==3.8.0
django-emailit==0.2.4
django-filer==2.0.2
django-filter==2.4.0
django-formtools==2.3
django-js-asset==1.2.2
django-meta==2.0.0
django-mptt==0.12.0
django-parler==2.2
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-sizefield==1.0.0
django-sortedm2m==3.0.2
django-tablib==3.2
django-taggit==1.4.0
django-taggit-autosuggest==0.3.8
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.4
django-widget-tweaks==1.4.8
djangocms-admin-style==2.0.2
djangocms-apphook-setup==0.4.1
djangocms-attributes-field==2.0.0
djangocms-blog==1.2.3
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-installer==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
djangorestframework==3.12.4
drf-yasg==1.20.0
easy-thumbnails==2.7.1
gunicorn==20.1.0
html5lib==1.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
kombu==5.1.0
lxml==4.6.3
MarkupSafe==2.0.1
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==21.0
Pillow==8.2.0
premailer==3.8.0
prompt-toolkit==3.0.19
pyotp==2.6.0
pyparsing==2.4.7
pytz==2021.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
tzlocal==2.1
Unidecode==1.1.2
uritemplate==3.0.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xmltodict==0.12.0
YURL==1.0.0
